fix(bindings): add "gprat" as first argument to HPX in start_hpx_wrapper if zero arguments
The first arguments given to HPX should hold the program name. Before
the commit, given zero arguments to the function start_hpx_wrapper, HPX
would be given "--hpx:threads=x" as the first argument for x CPU cores.
HPX does not apply the option but assumes that this is the program name.
Therefore, this commit regards this case of zero arguments to the
function and sets "gprat" as the first argument. Then, "--hpx:threads=x"
is used as the second argument and successfully sets the number of CPU
cores for the HPX runtime system to use.

Fixes #31