feat(core): add CUDA utility functions
When working with CUDA, there are some frequently used function
sequences that are wrapped in some functions in gpu/cuda_utils.cuh:
copy_to_device copies a host vector to the device and allocates its
device memory. The function create_cusolver_handle and destroy create
and destroy a cuSolver handle. Lastly, the free function takes a vector
of futures of double arrays (stored on the GPU) and frees its device
memory.